renamed trailer/cmd t0 cmd/trailer as per the go guidelines
