+++ cmd/trailer/utils.go
