+++ cmd/trailer/utils_test.go
